requirements.txt
torch
torchvision
pytorch-lightning
torchmetrics
wandb
hydra-core
hydra-colorlog
hydra-optuna-sweeper
black
flake8
isort
pre-commit
jupyterlab
jupyterlab-git
rich
tqdm
numpy
seaborn
opencv-python
timm